Asesoria crud ok

+++ src/main/java/cl/m5d12/servicio/AccidenteImple.java
@@ -15,9 +15,8 @@ public class AccidenteImple implements AccidenteServicio{
 	AccidenteRepository ar;
 	
 	@Override
-	public Accidente findAccidenteByid(int idAccidente) {
-
-		return ar.findOne(idAccidente); 
+	public Accidente findAccidenteByid(long idAccidente) {
+		return ar.findOne((int) idAccidente); 
 	}
 
 	@Override
