+++ src/main/java/cl/m5d12/servicio/AccidenteServicio.java
@@ -6,11 +6,12 @@ import cl.m5d12.dao.Accidente;
 
 public interface AccidenteServicio {
 	
-	Accidente findAccidenteByid(int idAccidente);
+	Accidente findAccidenteByid(long idAccidente);
 	List<Accidente> listaAccidente();
 	void addAccidente(Accidente acc);
 	void deleteAccidente(int idAccidente);
 	void updateAccidente(Accidente acc);
+
